Clean And BKP for Original Root SRC Project Files
